+++ app/helpers/node-by-type-total-time.js
@@ -1,0 +1,9 @@
+import { helper } from '@ember/component/helper';
+
+function nodeByTypeTotalTime(args) {
+  const [nodes] = args;
+
+  return nodes.map((node) => node.time).reduce((a, b) => a + b);
+}
+
+export default helper(nodeByTypeTotalTime);