+++ app/helpers/to-fixed.js
@@ -2,6 +2,7 @@ import { helper } from '@ember/component/helper';
 
 function toFixed(args) {
   let [number, to] = args;
+
   return Number(number).toFixed(to);
 }